+++ Muje/Muje/Models/ViewData/PostSuggestion.swift
@@ -1,0 +1,16 @@
+//
+//  PostSuggestion.swift
+//  Muje
+//
+//  Created by 조재훈 on 8/28/25.
+//
+
+import Foundation
+
+struct PostSuggestion: Codable, Identifiable {
+  let postId: String
+  let title: String
+  let organization: String
+  
+  var id: String { postId }
+}
